+++ src/main/java/uz/ilyoskhurozov/anyroute/util/CalculateReliability.java
@@ -4,8 +4,8 @@ import java.util.*;
 
 public class CalculateReliability {
 
-    public static float inModeVirtualChannel(Map<String, Float> routersRel, Map<String, Map<String, Float>> conRelTable, List<String> route){
-        float k = 1;
+    public static double inModeVirtualChannel(Map<String, Double> routersRel, Map<String, Map<String, Double>> conRelTable, List<String> route){
+        double k = 1;
 
         Iterator<String> routeIter = route.iterator();
 
@@ -23,16 +23,16 @@ public class CalculateReliability {
         return k;
     }
 
-    public static float inModeDatagram(Map<String, Float> routersRel, Map<String, Map<String, Float>> conRelTable, String begin, String end){
+    public static double inModeDatagram(Map<String, Double> routersRel, Map<String, Map<String, Double>> conRelTable, String begin, String end){
         List<String> routes = new ArrayList<>();
         findAllPaths(conRelTable, begin+";-;", begin, begin, end, routes);
 
-        float mid = 1;
+        double mid = 1;
         int k = 0;
         for (String route : routes) {
             String[] arr = route.split(";");
 
-            float sub = 1;
+            double sub = 1;
 
             for(int i = 1; i < arr.length-1; i++){
                 if (arr[i].equals("-")) {
@@ -47,7 +47,7 @@ public class CalculateReliability {
         return routersRel.get(begin) * (1 - mid) * routersRel.get(end);
     }
 
-    private static void findAllPaths(Map<String, Map<String, Float>> conRelTable, String route, String beginR, String curR, String endR, List<String> routes){
+    private static void findAllPaths(Map<String, Map<String, Double>> conRelTable, String route, String beginR, String curR, String endR, List<String> routes){
         conRelTable.get(curR).forEach((r, rel) -> {
             if (rel > 0 && !route.contains(r + ";") && !r.equals(beginR)) {
                 if (r.equals(endR)) {
